Add files via upload
This is the last major commit that I can do without the Firebase part of the project. Here are the changes:
- I added MORE files that you have to download for everything to work properly (HomeApiTask.kt), sorry about that.
- I got the search page working properly using some massive jank.
- I managed to get the home page favorites working using even more jank.

For the Firebase stuff, it would be best if you could return the latitude and longitude of the location that we want rather than the location name, but either one would work.